Move Question interface to separate file
