+++ src/objects.test.ts
@@ -1,7 +1,7 @@
+import { Question } from "./interfaces/question";
 import {
     makeBlankQuestion,
     isCorrect,
-    Question,
     isValid,
     toShortForm,
     toMarkdown,
